Non-detailed of Event Viewing
Added the event viewing, with all the components. The filter still need lots of work and the search system is working but it need some work to filter the page base on the search. Date is also showing but also need work on filtering the search. The program got all the events but from some reason it is stuck in the loading screen.